+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/questionnaire/AlertDialogViewModel.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.smartregister.fhircore.quest.ui.dialog
+package org.smartregister.fhircore.quest.ui.questionnaire
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
